scripts/python/instagram_scraper/requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.9.0
asyncio>=3.4.3
python-dotenv>=1.0.0

# Supabase integration
supabase>=2.3.0

# Scheduling and async
schedule>=1.2.0

# Data processing
pandas>=2.1.0

# Logging and monitoring
python-json-logger>=2.0.7

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional for advanced features
redis>=5.0.0  # For distributed queue management
influxdb>=5.3.0  # For metrics monitoring
boto3>=1.28.0  # For S3 backup support